Move top level remove top level logic into functions, add minor mode
This solves the long-standing issue where code was executed by just
loading the module, or by accessing the autoloads, which caused
erratic behavior when the package was installed by `package.el`.

This is a breaking change, since users must now explicitly enable the
theme switcher in their configuration.

The simplest way to do this is to add

```elisp
(moe-theme-switcher-mode 1)
```

to the configuration file.

Fixes #3
Fixes #31
Fixes #42
Fixes #68